requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0

# Documentos
python-docx==1.1.2
reportlab==4.2.2
python-pptx==1.0.2

# Grafos
networkx==3.3
matplotlib==3.9.2
requests==2.32.3

# UML (PlantUML vía HTTP)
httpx>=0.28.1

